+++ benchmarks/abc_app.ts
@@ -1,16 +1,16 @@
 import staticRoutes from "./static_routes.json";
 import paramsRoutes from "./params_routes.json";
-import { abc } from "../abc.ts";
+import { Application } from "../mod.ts";
 
-const Port = 8080;
+const port = 8080;
 
-const app = abc();
+const app = new Application();
 for (const r of staticRoutes) {
   app.any(r, (): string => r);
 }
 for (const r of paramsRoutes) {
   app.any(r, (): string => r);
 }
-app.start({ port: Port });
+app.start({ port });
 
-console.log(`server listening on ${Port}`);
+console.log(`server listening on http://localhost:${port}`);
