+++ deps.ts
@@ -2,9 +2,12 @@ import * as _path from "https://deno.land/std@v0.35.0/path/mod.ts";
 
 export {
   serve,
+  serveTLS,
   ServerRequest,
   Response,
-  Server
+  Server,
+  HTTPOptions,
+  HTTPSOptions
 } from "https://deno.land/std@v0.35.0/http/server.ts";
 export {
   Status,
