+++ docs/cors.md
@@ -10,7 +10,7 @@ const config: CORSConfig = {
   allowOrigins: ["https://a.com", "https://b.com", "https://c.com"],
   allowMethods: [HttpMethod.Get]
 };
-const app = abc();
+const app = new Application();
 app.use(cors(config));
 ```
