+++ docs/logger.md
@@ -5,10 +5,10 @@ Logger logs the information about each HTTP request.
 ### Usage
 
 ```ts
-import { abc } from "https://deno.land/x/abc/mod.ts";
+import { Application } from "https://deno.land/x/abc/mod.ts";
 import { logger } from "https://deno.land/x/abc/middleware/logger.ts";
 
-const app = abc();
+const app = new Application();
 app.use(logger());
 ```
