+++ docs/middleware.md
@@ -5,17 +5,15 @@ Middleware is a function which is called around the route handler. Middleware fu
 Let's start by implementing a simple middleware feature.
 
 ```ts
-function track(next: HandlerFunc) {
-  return function(c: Context) {
-    console.log(`request to ${c.path}`);
-    next(c);
-  };
-}
+const track: MiddlewareFunc = next => c => {
+  console.log(`request to ${c.path}`);
+  next(c);
+};
 ```
 
 ## Levels
 
-- Root Level: Root level middleware is executed before router processes the request. It can be registered via `abc().pre()`.
+- Root Level: Root level middleware is executed before router processes the request. It can be registered via `pre`.
 
 - Group Level: When creating a new group, you can register middleware just for that group.
 
@@ -23,7 +21,7 @@ function track(next: HandlerFunc) {
 
 ## Skipper
 
-There are cases when you would like to skip a middleware based on some conditions, for that each middleware has an option to define a function `skipper(c: Context): boolean`.
+There are cases when you would like to skip a middleware based on some conditions, for that each middleware has an option to define a function `skipper(c: IContext): boolean`.
 
 ```ts
 abc().use(
