+++ docs/static_files.md
@@ -1,13 +1,13 @@
 ## Static Files
 
-`abc.static()` registers a new route with path prefix to serve static files from the provided root directory. For example, a request to `/static/js/main.js` will fetch and serve `assets/js/main.js` file.
+`static` registers a new route with path prefix to serve static files from the provided root directory. For example, a request to `/static/js/main.js` will fetch and serve `assets/js/main.js` file.
 
 ```ts
-abc().static("/static", "assets");
+app.static("/static", "assets");
 ```
 
 `abc.file()` registers a new route with path to serve a static file.
 
 ```ts
-abc().file("/", "public/index.html");
+app.file("/", "public/index.html");
 ```
