+++ docs/table_of_contents.md
@@ -1,11 +1,8 @@
-English | [简体中文](./zh_CN/table_of_contents.md)
-
 ## Getting Started
 
 - [Getting Started](./getting_started.md)
 - [Router](./router.md)
 - [Middleware](./middleware.md)
-- [Data Binding](./data_binding.md)
 - [Static Files](./static_files.md)
 - [Exception Filter](./exception_filter.md)
