--- docs/zh_CN/table_of_contents.md
@@ -1,20 +1,0 @@
-[English](../table_of_contents.md) | 简体中文
-
-## 入门
-
-- [Getting Started 入门](./getting_started.md)
-- [Router 路由](./router.md)
-- [Middleware 中间件](./middleware.md)
-- [Data Binding 数据绑定](./data_binding.md)
-- [Static Files 静态资源](./static_files.md)
-- [Exception Filter 异常过滤](./exception_filter.md)
-
-## 中间件
-
-- [Logger 日志](./logger.md)
-- [CORS 跨域](./cors.md)
-
-## 风格指南
-
-- [Deno 风格指南](https://deno.land/std/style_guide.md)
-- [Abc 风格指南](./style_guide.md)
