+++ examples/01_cat_app/README.md
@@ -1,10 +1,8 @@
-- [cat_handler.ts](./cat_handler.ts)
-- [cat_router.ts](./cat_router.ts)
 - [cat.ts](./cat.ts)
-- [tsconfig.json](./tsconfig.json)
+- [main.ts](./main.ts)
 
 ### Usage
 
 ```shell
-deno --allow-net -c tsconfig.json cat_router.ts
+deno -A main.ts
 ```
