+++ examples/02_template/main.ts
@@ -1,7 +1,7 @@
-import { abc, Context } from "../../mod.ts";
+import { Application } from "../../mod.ts";
 import { renderFile } from "https://deno.land/x/dejs/mod.ts";
 
-const app = abc();
+const app = new Application();
 
 app.renderer = {
   render<T>(name: string, data: T): Promise<Deno.Reader> {
@@ -10,12 +10,9 @@ app.renderer = {
 };
 
 app
-  .get(
-    "/",
-    async (c: Context): Promise<void> => {
-      await c.render("./index.html", { name: "zhmushan" });
-    }
-  )
+  .get("/", async c => {
+    await c.render("./index.html", { name: "zhmushan" });
+  })
   .start({ port: 8080 });
 
-console.log(`server listening on 8080`);
+console.log(`server listening on http://localhost:8080`);
