+++ group.ts
@@ -1,19 +1,20 @@
-import { Abc, MiddlewareFunc, HandlerFunc, NotFoundHandler } from "./abc.ts";
+import { NotFoundHandler } from "./app.ts";
 import { path } from "./deps.ts";
+import { MiddlewareFunc, HandlerFunc, IApplication, IGroup } from "./types.ts";
 
-export class Group {
+export default class implements IGroup {
   prefix: string;
   middleware: MiddlewareFunc[];
-  abc: Abc;
+  app: IApplication;
 
-  constructor(options: GroupOptions) {
-    this.prefix = options.prefix || "";
-    this.abc = options.abc || ({} as Abc);
+  constructor(opts: { app: IApplication; prefix: string }) {
+    this.prefix = opts.prefix || "";
+    this.app = opts.app || ({} as IApplication);
 
     this.middleware = [];
   }
 
-  use(...m: MiddlewareFunc[]): Group {
+  use(...m: MiddlewareFunc[]): IGroup {
     this.middleware.push(...m);
     if (this.middleware.length !== 0) {
       this.any("", NotFoundHandler);
@@ -21,34 +22,34 @@ export class Group {
     return this;
   }
 
-  connect(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  connect(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("CONNECT", path, h, ...m);
   }
-  delete(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  delete(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("DELETE", path, h, ...m);
   }
-  get(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  get(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("GET", path, h, ...m);
   }
-  head(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  head(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("HEAD", path, h, ...m);
   }
-  options(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  options(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("OPTIONS", path, h, ...m);
   }
-  patch(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  patch(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("PATCH", path, h, ...m);
   }
-  post(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  post(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("POST", path, h, ...m);
   }
-  put(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  put(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("PUT", path, h, ...m);
   }
-  trace(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  trace(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     return this.add("TRACE", path, h, ...m);
   }
-  any(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): Group {
+  any(path: string, h: HandlerFunc, ...m: MiddlewareFunc[]): IGroup {
     const methods = [
       "CONNECT",
       "DELETE",
@@ -70,7 +71,7 @@ export class Group {
     path: string,
     h: HandlerFunc,
     ...m: MiddlewareFunc[]
-  ): Group {
+  ): IGroup {
     for (const method of methods) {
       this.add(method, path, h, ...m);
     }
@@ -81,8 +82,8 @@ export class Group {
     path: string,
     handler: HandlerFunc,
     ...middleware: MiddlewareFunc[]
-  ): Group {
-    this.abc.add(
+  ): IGroup {
+    this.app.add(
       method,
       this.prefix + path,
       handler,
@@ -92,28 +93,18 @@ export class Group {
     return this;
   }
 
-  static(prefix: string, root: string): Group {
-    this.abc.static(path.join(this.prefix, prefix), root);
+  static(prefix: string, root: string): IGroup {
+    this.app.static(path.join(this.prefix, prefix), root);
     return this;
   }
 
-  file(p: string, filepath: string, ...m: MiddlewareFunc[]): Group {
-    this.abc.file(path.join(this.prefix, p), filepath, ...m);
+  file(p: string, filepath: string, ...m: MiddlewareFunc[]): IGroup {
+    this.app.file(path.join(this.prefix, p), filepath, ...m);
     return this;
   }
 
-  group(prefix: string, ...m: MiddlewareFunc[]): Group {
-    const g = this.abc.group(this.prefix + prefix, ...this.middleware, ...m);
+  group(prefix: string, ...m: MiddlewareFunc[]): IGroup {
+    const g = this.app.group(this.prefix + prefix, ...this.middleware, ...m);
     return g;
   }
-}
-
-export interface GroupOptions {
-  abc: Abc;
-  prefix: string;
-}
-
-export function group(options: GroupOptions = {} as GroupOptions): Group {
-  const g = new Group(options);
-  return g;
 }
