+++ middleware/cors_test.ts
@@ -1,7 +1,7 @@
 import { assertEquals, runIfMain } from "../dev_deps.ts";
 import { cors } from "./cors.ts";
-import { Context } from "../context.ts";
 import { Header } from "../constants.ts";
+import { IContext } from "../types.ts";
 const { test } = Deno;
 
 test(function MiddlewareCORS(): void {
@@ -13,9 +13,8 @@ test(function MiddlewareCORS(): void {
     response: {
       headers
     }
-  } as Context;
-  cors()((c: Context): Context => c)(ctx);
-
+  } as IContext;
+  cors()(c => c)(ctx);
   assertEquals(headers.get(Header.Vary), Header.Origin);
   assertEquals(headers.get(Header.AccessControlAllowOrigin), "*");
 });
