--- tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "module": "esnext",
-    "resolveJsonModule": true,
-    "target": "esnext",
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true
-  }
-}
