--- tslint.json
@@ -1,14 +1,0 @@
-{
-  "rules": {
-    "typedef": [
-      true,
-      "call-signature",
-      "arrow-call-signature",
-      "property-declaration",
-      "parameter",
-      "arrow-parameter"
-    ],
-    "semicolon": true,
-    "quotemark": [true, "double"]
-  }
-}
