--- type.ts
@@ -1,3 +1,0 @@
-export interface Type<T> extends Function {
-  new(...args: any[]): T;
-}
